backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic-settings==2.5.2
pydantic[email]==2.9.2
jinja2==3.1.4
aiofiles==24.1.0
pillow==10.4.0
python-slugify==8.0.4
